requirements.txt
streamlit
streamlit-aggrid
streamlit-plotly-events
streamlit-option-menu
streamlit-folium
plotly
matplotlib
seaborn
obspy
numpy
scipy
pandas
h5py
requests
python-dotenv
pyproj
agno
duckduckgo-search
ollama
httpx
litellm
openai
anthropic
google-generativeai
reportlab
openpyxl
fpdf2
pytest
pytest-cov
black
flake8
mypy
pyinstaller
